[SDESK-4224][SDESK-4510][SDESK-4513] (fix): Don't unmount the PopupEditor when action modal is shown Uses React.Portal for the PopupEditor and 'display: none' on the editor when an action modal is shown
