+++ client/components/Main/ItemEditor/Editor.jsx
@@ -179,13 +179,12 @@ export class EditorComponent extends React.Component {
                     itemId,
                     itemType,
                     addNewsItemToPlanning,
-                    inModalView,
                 } = this.props;
                 const {dirty, errorMessages, initialValues} = this.state;
 
                 this.setState({submitting: true});
 
-                const updateStates = !addNewsItemToPlanning && !inModalView;
+                const updateStates = !addNewsItemToPlanning;
 
                 if (!dirty) {
                     this.onCancel();
@@ -239,9 +238,7 @@ export class EditorComponent extends React.Component {
         if (updateStates) {
             this.setState({submitting: false});
 
-            if (!this.props.inModalView &&
-                (this.tearDownRequired || !isExistingItem(this.state.initialValues))
-            ) {
+            if (this.tearDownRequired || !isExistingItem(this.state.initialValues)) {
                 this.tearDownEditorState();
             }
         }
